+++ fiftyfive-wicket-core/src/test/java/fiftyfive/wicket/data/DtoDataProviderTest.java
@@ -15,12 +15,36 @@
  */
 package fiftyfive.wicket.data;
 
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
 import fiftyfive.wicket.BaseWicketTest;
 import org.junit.Assert;
 import org.junit.Test;
 
 public class DtoDataProviderTest extends BaseWicketTest
 {
+    /**
+     * Verify that if we call iterator() explicitly that we get the desired
+     * results, instead of cached ones.
+     */
+    @Test
+    public void testExplicitIterate() throws Exception
+    {
+        DtoDataProviderTestPage page = assertRenderAndClickNext(0);
+        BeanResultProvider provider = page.getProvider();
+        
+        Iterator<? extends Bean> iter = provider.iterator(0, provider.size());
+        List<Bean> list = new ArrayList<Bean>(100);
+        while(iter.hasNext())
+        {
+            list.add(iter.next());
+        }
+        
+        Assert.assertEquals(100, list.size());
+        Assert.assertEquals(2, provider.getLoadCount());
+    }
+    
     @Test
     public void testRender() throws Exception
     {
@@ -39,7 +63,7 @@ public class DtoDataProviderTest extends BaseWicketTest
         assertRenderAndClickNext(2);
     }
     
-    private void assertRenderAndClickNext(int clicks) throws Exception
+    private DtoDataProviderTestPage assertRenderAndClickNext(int clicks) throws Exception
     {
         DtoDataProviderTestPage page = new DtoDataProviderTestPage();
         this.tester.startPage(page);
@@ -56,5 +80,7 @@ public class DtoDataProviderTest extends BaseWicketTest
         
         // Data should have been loaded only once per render!
         Assert.assertEquals(1 + clicks, page.getLoadCount());
+        
+        return page;
     }
 }